+++ google/cloud/bigtable/_helpers.py
@@ -13,6 +13,15 @@
 #
 from __future__ import annotations
 
+from inspect import iscoroutinefunction
+from typing import Callable, Any
+
+from google.api_core import exceptions as core_exceptions
+from google.cloud.bigtable.exceptions import RetryExceptionGroup
+"""
+Helper functions used in various places in the library.
+"""
+
 
 def _make_metadata(
     table_name: str | None, app_profile_id: str | None
@@ -27,3 +36,51 @@ def _make_metadata(
         params.append(f"app_profile_id={app_profile_id}")
     params_str = ",".join(params)
     return [("x-goog-request-params", params_str)]
+
+
+def _convert_retry_deadline(
+    func: Callable[..., Any],
+    timeout_value: float | None = None,
+    retry_errors: list[Exception] | None = None,
+):
+    """
+    Decorator to convert RetryErrors raised by api_core.retry into
+    DeadlineExceeded exceptions, indicating that the underlying retries have
+    exhaused the timeout value.
+    Optionally attaches a RetryExceptionGroup to the DeadlineExceeded.__cause__,
+    detailing the failed exceptions associated with each retry.
+
+    Supports both sync and async function wrapping.
+
+    Args:
+      - func: The function to decorate
+      - timeout_value: The timeout value to display in the DeadlineExceeded error message
+      - retry_errors: An optional list of exceptions to attach as a RetryExceptionGroup to the DeadlineExceeded.__cause__
+    """
+    timeout_str = f" of {timeout_value:.1f}s" if timeout_value is not None else ""
+    error_str = f"operation_timeout{timeout_str} exceeded"
+
+    def handle_error():
+        new_exc = core_exceptions.DeadlineExceeded(
+            error_str,
+        )
+        source_exc = None
+        if retry_errors:
+            source_exc = RetryExceptionGroup(retry_errors)
+        new_exc.__cause__ = source_exc
+        raise new_exc from source_exc
+
+    # separate wrappers for async and sync functions
+    async def wrapper_async(*args, **kwargs):
+        try:
+            return await func(*args, **kwargs)
+        except core_exceptions.RetryError:
+            handle_error()
+
+    def wrapper(*args, **kwargs):
+        try:
+            return func(*args, **kwargs)
+        except core_exceptions.RetryError:
+            handle_error()
+
+    return wrapper_async if iscoroutinefunction(func) else wrapper
